src/nlp_text_summary_generator/utils/common.py
import os
from box.exceptions import BoxValueError
import yaml
from nlp_text_summary_generator.logging import logger
from ensure import ensure_annotations
from pathlib import Path
from typing import Any