requirements.txt
absl-py==2.1.0
aiohappyeyeballs @ file:///C:/b/abs_b505trsapr/croot/aiohappyeyeballs_1725434036096/work
aiohttp @ file:///C:/b/abs_13j6efxjb7/croot/aiohttp_1725529348885/work
aiosignal @ file:///tmp/build/80754af9/aiosignal_1637843061372/work
annotated-types==0.7.0
anyio==4.5.2
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
async-timeout @ file:///C:/b/abs_c8fgiuixkq/croot/async-timeout_1703097556097/work
attrs @ file:///C:/b/abs_24xj44kobp/croot/attrs_1729089434205/work
babel==2.16.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
boto3==1.35.49
botocore==1.35.49
Bottleneck @ file:///C:/b/abs_f05kqh7yvj/croot/bottleneck_1707864273291/work
Brotli==1.1.0
certifi @ file:///C:/b/abs_1fw_exq1si/croot/certifi_1725551736618/work/certifi
cffi==1.17.1
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
click==8.1.7
colorama @ file:///C:/b/abs_a9ozq0l032/croot/colorama_1672387194846/work
comm==0.2.2
contourpy==1.1.1
cycler==0.12.1
datasets @ file:///C:/b/abs_62zngb45t8/croot/datasets_1716911669805/work
debugpy==1.8.7
decorator==5.1.1
defusedxml==0.7.1
dill @ file:///C:/b/abs_f79eg27d2q/croot/dill_1715094735295/work
ensure==1.0.4
exceptiongroup==1.2.2
executing==2.1.0
fastapi==0.115.4
fastjsonschema==2.20.0
filelock @ file:///C:/b/abs_f2gie28u58/croot/filelock_1700591233643/work
fonttools==4.54.1
fqdn==1.5.1
frozenlist @ file:///C:/b/abs_d8e__s1ys3/croot/frozenlist_1698702612014/work
fsspec @ file:///C:/b/abs_d5jzmndkba/croot/fsspec_1714461594761/work
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
huggingface_hub @ file:///C:/b/abs_6f4hr2sio5/croot/huggingface_hub_1724853960198/work
idna @ file:///C:/b/abs_aad84bnnw5/croot/idna_1714398896795/work
importlib_metadata==8.5.0
importlib_resources==6.4.5
inflate64==1.0.0
ipykernel==6.29.5
ipython==8.12.3
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
kiwisolver==1.4.7
lxml==5.3.0
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.7
mistune==3.0.2
mkl-fft @ file:///C:/b/abs_19i1y8ykas/croot/mkl_fft_1695058226480/work
mkl-random @ file:///C:/b/abs_edwkj1_o69/croot/mkl_random_1695059866750/work
mkl-service==2.4.0
multidict @ file:///C:/b/abs_44ido987fv/croot/multidict_1701097803486/work
multiprocess @ file:///C:/b/abs_43lxksn2oj/croot/multiprocess_1692295065414/work
multivolumefile==0.2.3
mypy-boto3-s3==1.35.46
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
-e git+https://github.com/itadityaa/TextSummarization.git@93ce27a896dc61a50b4493c82bfd67a67d0aed0f#egg=nlp_text_summary_generator
nltk==3.9.1
notebook==7.2.2
notebook_shim==0.2.4
numexpr @ file:///C:/b/abs_afm0oewmmt/croot/numexpr_1683221839116/work
numpy @ file:///C:/Users/dev-admin/mkl/numpy_and_numpy_base_1682982345978/work
overrides==7.7.0
packaging @ file:///C:/b/abs_c3vlh0z4jw/croot/packaging_1720101866539/work
pandas @ file:///C:/miniconda3/conda-bld/pandas_1692299636855/work
pandocfilters==1.5.1
parso==0.8.4
pickleshare==0.7.5
pillow==10.4.0
pkgutil_resolve_name==1.3.10
platformdirs==4.3.6
portalocker==2.10.1
prometheus_client==0.21.0
prompt_toolkit==3.0.48
protobuf==5.28.3
psutil==6.1.0
pure_eval==0.2.3
py7zr==0.22.0
pyarrow @ file:///C:/b/abs_cbc4sg_1e0/croot/pyarrow_1721664257308/work/python
pybcj==1.0.2
pycparser==2.22
pycryptodomex==3.21.0
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pyparsing==3.1.4
pyppmd==1.1.0
PySocks @ file:///C:/ci/pysocks_1605287845585/work
python-box==7.2.0
python-dateutil @ file:///C:/b/abs_3au_koqnbs/croot/python-dateutil_1716495777160/work
python-json-logger==2.0.7
pytz @ file:///C:/b/abs_6ap4tsz1ox/croot/pytz_1713974360290/work
pywin32==308
pywinpty==2.0.14
PyYAML @ file:///C:/b/abs_14xkfs39bx/croot/pyyaml_1728657968772/work
pyzmq==26.2.0
pyzstd==0.16.2
referencing==0.35.1
regex @ file:///C:/b/abs_5cm86yjgo3/croot/regex_1726670543261/work
requests @ file:///C:/b/abs_9frifg92q2/croot/requests_1721410901096/work
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rouge_score==0.1.2
rpds-py==0.20.0
s3transfer==0.10.3
sacrebleu==2.4.3
safetensors @ file:///C:/b/abs_f8a0oan4c9/croot/safetensors_1729663460267/work
Send2Trash==1.8.3
sentencepiece==0.2.0
six @ file:///tmp/build/80754af9/six_1644875935023/work
sniffio==1.3.1
soupsieve==2.6
stack-data==0.6.3
starlette==0.41.2
tabulate==0.9.0
terminado==0.18.1
texttable==1.7.0
tinycss2==1.4.0
tokenizers @ file:///C:/b/abs_dbyirlxdi3/croot/tokenizers_1729696708579/work
tomli==2.0.2
tornado==6.4.1
tqdm @ file:///C:/b/abs_77wju137gk/croot/tqdm_1724853945787/work
traitlets==5.14.3
transformers @ file:///C:/b/abs_2bgndud1b2/croot/transformers_1729767177936/work
types-python-dateutil==2.9.0.20241003
typing_extensions @ file:///C:/b/abs_0as9mdbkfl/croot/typing_extensions_1715268906610/work
tzdata @ file:///croot/python-tzdata_1690578112552/work
uri-template==1.3.0
urllib3==1.26.20
uvicorn==0.32.0
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
win-inet-pton @ file:///C:/ci/win_inet_pton_1605306167264/work
xxhash @ file:///C:/b/abs_e59ry5bslh/croot/python-xxhash_1667919514787/work
yarl @ file:///C:/b/abs_00rcuicx5k/croot/yarl_1726015884215/work
zipp==3.20.2
-e .